Split FormattableTextArea into more files
