+++ src/text/format.cpp
@@ -29,12 +29,6 @@ namespace {
     }
 }
 
-//! Gets a new QTextCharFormat containing all the formats found in the text
-//! cursor's selection, or at its position if no text is selected.
-//! This function only returns the value if the entire range is styled.
-//! For instance, if the cursor selects a range where half the range is
-//! italicized and half the range is not, the resulting format will have
-//! fontItalics() set to false.
 const QTextCharFormat format::getMergedCharFormat(const QTextCursor& textCursor)
 {
     const int start = textCursor.selectionStart();
